+++ apps/database-new/components/Chat/Chat.tsx
@@ -10,7 +10,7 @@ import { useMessagesQuery } from '@/data/messages-query'
 import { AssistantMessage, UserMessage } from '@/lib/types'
 import BottomMarker from './BottomMarker'
 import UserChat from './UserChat'
-import { ChatInputAtom } from './ChatInputAtom'
+import { ChatInputAtom } from './ChatInput'
 
 export const Chat = () => {
   const router = useRouter()
